+++ notifications/urls.py
@@ -9,4 +9,3 @@ router = DefaultRouter()
 router.register(r"notifications", NotificationViewSet, basename="notifications")
 
 urlpatterns = router.urls
-