+++ procurement_tool/asgi.py
@@ -7,10 +7,19 @@ For more information on this file, see
 https://docs.djangoproject.com/en/5.0/howto/deployment/asgi/
 """
 
+"""
+ASGI config for procurement_tool project.
+
+It exposes the ASGI callable as a module-level variable named ``application``.
+
+For more information on this file, see
+https://docs.djangoproject.com/en/5.0/howto/deployment/asgi/
+"""
+
 import os
 
 from django.core.asgi import get_asgi_application
 
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'procurement_tool.settings')
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "procurement_tool.settings")
 
 application = get_asgi_application()